Add data logic from connection to data_manager
