+++ server.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, redirect, url_for, request
-import connection
+import data_manager
 
 app = Flask(__name__)
 
@@ -7,7 +7,7 @@ app = Flask(__name__)
 @app.route('/')
 @app.route('/list')
 def route_list():
-    stored_questions = connection.formatted_file()
+    stored_questions = data_manager.formatted_file()
     return render_template('list.html', questions=stored_questions)
